Adds support to SMF\TimeInterval for all \DateInterval methods
This is tricker than one might expect, because of some very quirky behaviour in \DateInterval and how it handles its properties. In particular, calling parent::__construct() from a child class will clobber certain properties values and make them read-only. But not calling the parent constructor means that we must implement all of the parent's methods and properties in the child class.
Signed-off-by: Jon Stovell <jonstovell@gmail.com>